@@ -1,10 +1,15 @@
 class CommentsController < ApplicationController
-	def create
-		@post = Post.find(params[:post_id])
-		@comment = @post.comments.create!(params[:comment])
+  def create
+    @post = Post.find(params[:post_id])
+    @comment = @post.comments.create!(comment_params)
     respond_to do |format|
       format.html { redirect_to @post }
       format.js
     end
-	end
+  end
+
+  private
+    def comment_params
+      params.require(:comment).permit(:body)
+    end
 end
