@@ -1,12 +1,6 @@
 class AdminUser < ActiveRecord::Base
-  devise :database_authenticatable,
-    :recoverable, :rememberable, :trackable, :validatable
-
-  attr_accessible :email, :password, :password_confirmation, :remember_me
-
-  has_many :posts
-
-  def name
-    email
-  end
+  # Include default devise modules. Others available are:
+  # :confirmable, :lockable, :timeoutable and :omniauthable
+  devise :database_authenticatable, 
+         :recoverable, :rememberable, :trackable, :validatable
 end
